Entity Framework altyapısı genel olarak hazırlandı ve tamamlandı.

+++ Entities/Concrete/Car.cs
@@ -13,7 +13,7 @@ namespace Entities.Concrete
 
         public int ModelYear { get; set; }
 
-        public decimal DailyPrice { get; set; }
+        public int DailyPrice { get; set; }
 
         public string Description { get; set; }
     }
